backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
celery==5.3.4
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
ffmpeg-python==0.2.0
Pillow==10.1.0
python-decouple==3.8
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1